move blockchain definition to separate file
